refactor: move shared enums
it's creating problems with `--build-optimizer`
