+++ src/app/shared/models/descriptive-statistic.enum.ts
@@ -1,0 +1,13 @@
+export enum DescriptiveStatistic {
+  AVERAGE,
+  COUNT,
+  MAXIMUM,
+  MEDIAN,
+  MINIMUM,
+  SUM,
+  INTERQUARTILE_RANGE,
+  LOWER_QUARTILE,
+  UPPER_QUARTILE,
+  QUARTILES,
+  RECEIVED_MESSAGES
+}
