+++ src/app/shared/models/time-interval.enum.ts
@@ -1,0 +1,9 @@
+export enum TimeInterval {
+  TEN_SECOND = 10000,
+  THIRTY_SECOND = 30000,
+  ONE_MIN = 60000,
+  TEN_MIN = 600000,
+  ONE_HOUR = 3600000,
+  ONE_DAY = 86400000,
+  ONE_WEEK = 604800000
+}
