+++ src/app/shared/store/sensors/sensors.model.ts
@@ -27,30 +27,6 @@ export enum ChartType {
   bar = 'bar'
 }
 
-export enum TimeInterval {
-  TEN_SECOND = 10000,
-  THIRTY_SECOND = 30000,
-  ONE_MIN = 60000,
-  TEN_MIN = 600000,
-  ONE_HOUR = 3600000,
-  ONE_DAY = 86400000,
-  ONE_WEEK = 604800000
-}
-
-export enum DescriptiveStatistic {
-  AVERAGE,
-  COUNT,
-  MAXIMUM,
-  MEDIAN,
-  MINIMUM,
-  SUM,
-  INTERQUARTILE_RANGE,
-  LOWER_QUARTILE,
-  UPPER_QUARTILE,
-  QUARTILES,
-  RECEIVED_MESSAGES
-}
-
 export enum SensorType {
   ACCELEROMETER,
   BATTERY,
