+++ src/app/shared/testing/mocks/mock-compliance-data.ts
@@ -1,5 +1,5 @@
 import { ChartData } from '../../models/chart-data.model'
-import { TimeInterval } from '../../store/sensors/sensors.model'
+import { TimeInterval } from '../../models/time-interval.enum'
 import { getTime } from '../../utils/get-time'
 
 export const MockComplianceKeys = [
